Data.Thorn.Fmap

Description

The module Data.Thorn.Fmap

Documentation

autofmap :: TypeQ -> ExpQ

autofmap t generates the fmap of the type t. Quite surprisingly, it still works for any arity, co/contra/free/fixed-variances, type synonyms, and mutual recursions.

type Nuf x y = y -> x
type a :<- b = Nuf a b
nuf = $(autofmap [t|(:<-)|]) chr ord (+1) c

data List a = Nil | Cons a (List a) deriving Show
golist 0 = Nil
golist n = Cons n (golist (n-1))
list = $(autofmap $[t|List|]) (+1) (golist 10)

data Rose a = Rose a (Forest a) deriving Show
data Forest a = Forest [Rose a] deriving Show
gorose n = Rose n (Forest (replicate n (gorose (n-1))))
rose = $(autofmap $[t|Rose|]) (+1) (gorose 3)

data Variance

Variance is a variance of a parameter of a functor. Co means covariance, one of a normal functor, Contra means contravariance, a dual of covariance, Free means to be able to satisfy either covariance or contravariance, Fixed means to have to satisfy both covariance and contravariance. And v1 mappend v2 means to have to satisfy both v1 and v2.

Constructors

Co

Contra

Free

Fixed

autovariance :: TypeQ -> Q [Variance]

autovariance t provides you the variances.