Data.Thorn.Fold

Description

The module Data.Thorn.Fold.

Folding and Unfolding

Thorn generates folds and unfolds from various kinds of recursive datatypes, including mutually recursive ones.

unfixdata

Arguments

:: TypeQ
t, recursive datatype

-> String
s, name of the datatype to be declared

-> (String -> String)
f, how to convert the name of data constructors

-> [Name]
ds, derivings

-> DecsQ
declaration of a nonrecursive datatype whose fixpoint is t

unfixdata t n f ds provides a declaration of a nonrecursive datatype whose fixpoint is the recursive type t, with a deriving declaration with names ds.

autofold

Arguments

:: TypeQ
u, nonrecursive datatype

-> TypeQ
t, fixpoint of u

-> ExpQ
folding function with a type (u x0 .. xn a -> a) -> (t x0 .. xn -> a)

autofold u t provides a folding function for the recursive type t.

autofoldtype :: TypeQ -> TypeQ -> TypeQ

autofoldtype u t provides the type of $(autofold u t), that is, (u x0 .. xn a -> a) -> (t x0 .. xn -> a).

autofolddec :: String -> TypeQ -> TypeQ -> DecsQ

autofolddec s u t provides a declaration of a folding function for the recursive type t with the name s, with a type signature.

autounfold

Arguments

:: TypeQ
u, nonrecursive datatype

-> TypeQ
t, fixpoint of u

-> ExpQ
unfolding function with a type (a -> u x0 .. xn a) -> (a -> t x0 .. xn)

autounfold t provides an unfolding function for the recursive type t.

autounfoldtype :: TypeQ -> TypeQ -> TypeQ

autounfoldtype u t provides the type of $(autounfold u t), that is, (a -> u x0 .. xn a) -> (a -> t x0 .. xn).

autounfolddec :: String -> TypeQ -> TypeQ -> DecsQ

autounfolddec s u t provides a declaration of an unfolding function for the recursive type t with the name s, with a type signature.

Mutual Recursion

unfixdataMutual

Arguments

:: [(TypeQ, String, String -> String, [Name])]
[(t0,s0,f0,ds0), ...]; recursive datatype, name of the datatype to be declared, how to convert the name of data constructors, and derivings

-> DecsQ
declarations of datatypes u0, u1, u2, ..., whose fixpoints are t0, t1, t2, ... respectively

Mutually recursive version of unfixdata. Note that

unfixdata t s f ds = unfixdataMutual [(t,s,f,ds)]

autofoldMutual

Arguments

:: [(TypeQ, TypeQ)]
[(u0,t0), .., (un,tn)]; ui is an nonrecursive datatype and ti is a fixpoint of ui

-> Int
k, index

-> ExpQ
folding function with a type (u0 x0 .. xm a0 .. an -> a0) -> .. -> (un x0 .. xm a0 .. an -> an) -> (tk x0 .. xm -> ak)

autofoldMutual uts k provides a folding function for the mutually recursive type tk.

autofoldtypeMutual :: [(TypeQ, TypeQ)] -> Int -> TypeQ

autofoldtypeMutual uts k provides the type of $(autofoldMutual uts k), that is, (u0 x0 .. xm a0 .. an -> a0) -> .. -> (un x0 .. xm a0 .. an -> an) -> (tk x0 .. xm -> ak).

autofolddecMutual :: String -> [(TypeQ, TypeQ)] -> Int -> DecsQ

autofolddecMutual s uts k provides a declaration of a folding function for the mutually recursive type tk with the name s, with a type signature.

autounfoldMutual

Arguments

:: [(TypeQ, TypeQ)]
[(u0,t0), .., (un,tn)]; ui is an nonrecursive datatype and ti is a fixpoint of ui

-> Int
k, index

-> ExpQ
unfolding function with a type (a0 -> u0 x0 .. xm a0 .. an) -> .. -> (an -> un x0 .. xm a0 .. an) -> (ak -> tk x0 .. xm)

autounfoldMutual uts k provides an unfolding function for the mutually recursive type tk.

autounfoldtypeMutual :: [(TypeQ, TypeQ)] -> Int -> TypeQ

autounfoldtypeMutual uts k provides the type of $(autounfoldMutual uts k), that is, (a0 -> u0 x0 .. xm a0 .. an) -> .. -> (an -> un x0 .. xm a0 .. an) -> (ak -> tk x0 .. xm).

autounfolddecMutual :: String -> [(TypeQ, TypeQ)] -> Int -> DecsQ

autofolddecMutual s uts k provides a declaration of an unfolding function for the mutually recursive type tk with the name s, with a type signature.

Helper Function

modifynameunfix :: String -> String

Use this function to designate how to convert the name of data constructors for unfixdata.

 modifynameunfix "Hello" == "UfHello"
 modifynameunfix ":***" == ":&***"

Note that

modifynameunfix == modifyname ("Uf","") ("&","")

Primitive Functions

autoin

Arguments

:: TypeQ
u, nonrecursive datatype

-> TypeQ
t, fixpoint of u

-> ExpQ
function with a type u x0 .. xn t -> t x0 .. xn

autoout

Arguments

:: TypeQ
u, nonrecursive datatype

-> TypeQ
t, fixpoint of u

-> ExpQ
function with a type t x0 .. xn -> u x0 .. xn t

autohylo

Arguments

:: TypeQ
u, nonrecursive datatype

-> ExpQ
function with a type (a -> u x0 .. xn a) -> (u x0 .. xn b -> b) -> (a -> b)

autoinMutual

Arguments

:: [(TypeQ, TypeQ)]
[(u0,t0), .., (un,tn)]; ui is an nonrecursive datatype and ti is a fixpoint of ui

-> Int
k, index

-> ExpQ
function with a type uk x0 .. xm t0 .. tn -> tk x0 .. xm

Mutually recursive version of autoin.

autooutMutual

Arguments

:: [(TypeQ, TypeQ)]
[(u0,t0), .., (un,tn)]; ui is an nonrecursive datatype and ti is a fixpoint of ui

-> Int
k, index

-> ExpQ
function with a type tk x0 .. xm -> uk x0 .. xm t0 .. tn

Mutually recursive version of autoout.

autohyloMutual

Arguments

:: [TypeQ]
[u0, .., un]; ui is an nonrecursive datatype

-> Int
k, index

-> ExpQ
function with a type (a0 -> u0 x0 .. xm a0 .. an) -> .. -> (an -> un x0 .. xm a0 .. an) -> (u0 x0 .. xm b0 .. bn -> b0) -> (un x0 .. xm b0 .. bn -> bn) -> (ak -> bk)

Mutually recursive version of autohylo.

Examples

Basic

 oh

Mutual Recursion

 oh